requirements.txt
asgiref==3.9.2
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.1.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pytest==8.4.2
pytest-django==4.11.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
